test(Login): add integration tests for login route
